requirements.txt
sparkfun-qwiic-rfid>=2.0.0